Modify some Apis

+++ code/server/classes/InternalOrder/InternalOrderAPI.js
@@ -115,7 +115,29 @@ module.exports = function (app) {
         const customerID = req.body.customerId;
         if (date == undefined || products == undefined || customerID == undefined)
             return res.status(422).end();
+        var d1 = /^\d{4}\/(0?[1-9]|1[012])\/(0?[1-9]|[12][0-9]|3[01])$/;
+        var d2 = /^\d{4}\/(0?[1-9]|1[012])\/(0?[1-9]|[12][0-9]|3[01]) ([0-1]?[0-9]|2[0-3]):[0-5][0-9]$/;
+
+        if(!d1.test(date) && !d2.test(date)){
+            return res.status(422).end();
+        }
         
+        let user = await InternalOrderDao.getUserByID(customerID);
+        if(user == undefined) {
+            return res.status(422).end();
+        }
+
+        var valid=true;
+        products.forEach(async function(e){
+            let a = await InternalOrderDao.searchProductForSkuID(e.SKUId);
+            if(a == undefined || e.description == undefined || e.price == undefined || e.qty == undefined) {
+                valid=false;
+            }
+        } )
+        if(!valid){
+            return res.status(422).end();
+        }
+
         const id = await InternalOrderDao.storeInternalOrder({ date: date, state: 0, customerID: customerID });
     
         products.forEach(async function (e) {
@@ -124,7 +146,7 @@ module.exports = function (app) {
         });
 
 
-        return res.end();
+        return res.status(201).end();
     });
 
 
@@ -134,6 +156,11 @@ module.exports = function (app) {
          const id = req.params.id;
          if (id == undefined || newState == undefined)
              return res.status(422).end();
+        
+
+        if(newState!= 'COMPLETED' || newState!= 'ACCEPTED' || newState!= 'ISSUED' || newState!= 'REFUSED' || newState!= 'CANCELED')
+            return res.status(422).end();
+
         let order = await InternalOrderDao.getInternalOrderByID(id);
         if(order==undefined){
             return res.status(404).end();
@@ -146,7 +173,7 @@ module.exports = function (app) {
              });
          }
 
-         return res.sendStatus(200);
+         return res.status(200).end();
      });
 
     app.delete('/api/internalOrders/:id', async function (req, res) {
@@ -154,12 +181,17 @@ module.exports = function (app) {
         const id = req.params.id;
         if (id == undefined)
             return res.status(422).end();
+        
+
+        let internalOrder = await InternalOrderDao.getInternalOrderByID(id);
+        if (internalOrder == undefined)
+            return res.status(422).end();
 
         const result = await InternalOrderDao.deleteInternalOrder(id);
         if(!result)
-            return res.sendStatus(503);
+            return res.status(503).end();
 
-        return res.sendStatus(204);
+        return res.status(204).end();
     });
